+++ src/main/java/com/lokoko/domain/productBrand/api/dto/ProductBrandInfoProjection.java
@@ -1,4 +1,4 @@
-package com.lokoko.domain.productBrand.api.dto.response;
+package com.lokoko.domain.productBrand.api.dto;
 
 public record ProductBrandInfoProjection(
 	String productBrandName,
